Move UnrealURL to its own dedicated file
